Add @preparable, decouple prepare/[Multi]Injector
Since the concept of 'preparable component' is not specifically tied to
using a Nexus.Flux instance, there is now a dedicated `@preparable`
decorators which can be applied to any React.Component class to define how
to prepare for the rendering of the component, given its props. While it
can depend on props, this method is however static, as it can not rely on
eg. state of the component.

Accordingly, prepare now looks for the method added using `@preparable`
instead of specifically looking for descendant classes of Injector or
MultiInjector. Instead, Injector and MultiInjector have been decorated
with `@preparable` to produce the exact same behaviour as before.