Replace inject with multiInject
Old inject was practically useless, being a mere special case of
multiInject.

New `@inject` getInjected (former getBinding) callback now receives (nexus, props) instead of
(props, nexus), since nexus is always used and props not always.

The Flux.Binding type (ands its derived types, HTTPFlux.Binding and
LocalFlux.Binding) now enforce a contract: a value in the map returned by
getInjected describes a binding if and only if it is an instance of
Flux.Binding. Other values are passed directly to the children.
This allows to pass arbitrary expressions using props or nexus as derived
props to the children, effectively removing the need for either
getNexusOf (which was accordingly deleted) or react-transform-props.

To dispatch an action from the component, you must therefore decorate it
with `@inject`, and explicitly inject the reference to a flux as props
(see the User.jsx example in src/__tests__/fixtures/components).