Change prev/next bindings behavior
When an update occurs in `componentWillReceiveProps` React lifecycle
of `stores` decorator, new bindings and old bindings are identified.

For new added bindings:
- state is updated with associated value of store.
- component is registered as an observer.

For old removed bindings:
- component is unregistered from observers.

As you can notice, when old bindings are removed, state is not updated.
This means that the decorated component will always receive the previous
value of store.

We believe that it would be more appropriate to clean the related state
value.
We propose to set the value to `void 0`.