Make prepare produce a simplified react elements tree for renderToString
There was an issue with react nexus, to prepare the app and satisfy its
dependencies, it was required to walk through the whole elements tree and
call the render functions of each element.
Then, with the stores populated, the app was given to
`React.renderToString` which also needs to walk the whole elements tree
and execute each render function to build the string of html.

With this modification, during the prepare phase, a new element tree is
built using the data gathered through the elements tree walk, but only
with the minimal data required for `React.renderToString` to build the
exact same string of html as the one that would be produced by
processing the original app. This is much faster because there is no
render function to call anymore (except for `react-empty` tags).